% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{darker}
\alias{darker}
\title{darker}
\usage{
darker(col, amount=150)
}
\arguments{
\item{amount}{(integer, [0-255]) The fixed amount to add to each RGB channel (Defaults to 150).}

\item{color}{(color) The original color}
}
\value{
A darker color
}
\description{
Given a color, return a darker one
}
\details{
Very simple utility function to create darker colors. Given a color, it
transforms it to rgb space, adds a set amount to all chanels and transforms
it back to a color.
}
\examples{
 
darker("red")
darker("#333333")
 
}
\seealso{
\code{\link{lighter}}
}

