% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getColorSchemas}
\alias{getColorSchemas}
\title{getColorSchemas}
\usage{
getColorSchemas()
}
\value{
A list with the color schemas included in karyoploteR for cytobands,
variants, horizons...
}
\description{
Return a structure with the color schemas included in karyoploteR
}
\examples{
 
getColorSchemas()
 
}
