% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultPlotParams.R
\name{getDefaultPlotParams}
\alias{getDefaultPlotParams}
\title{getDefaultParameters}
\usage{
getDefaultPlotParams(plot.type)
}
\description{
Returns the default parameter for the given plot.type
}
\details{
TODO
}
\examples{

 
}
\seealso{
\code{\link{plotKaryotype}}, \code{\link{kpLines}}, \code{\link{kpPoints}}, \code{\link{kpPlotRegions}}
}

