% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{is.color}
\alias{is.color}
\title{is.color}
\usage{
is.color(x)
}
\arguments{
\item{x}{The element to test}
}
\value{
TRUE is x is a valid color, FALSE otherwise
}
\description{
Test if something is a valid color
}
\details{
This function tests if something is a valid color. Returns TRUE or FALSE. 
The function is vectorised.
}
\examples{
 
is.color("red")
is.color("#333333")
is.color(NA)
is.color(NULL)
is.color("not_a_color")
is.color(3)

is.color(c("not_a_color", "red", 3, "#FF0000"))
 
}
