\name{karyoploteR-package}
\alias{karyoploteR-package}
\alias{karyoploteR}
\docType{package}
\title{
\packageTitle{karyoploteR}
}
\description{
\packageDescription{karyoploteR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{karyoploteR}
\packageIndices{karyoploteR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{karyoploteR}

Maintainer: \packageMaintainer{karyoploteR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
