% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpHeatmap.R
\name{kpHeatmap}
\alias{kpHeatmap}
\title{kpHeatmap}
\usage{
kpHeatmap(karyoplot, data=NULL, chr=NULL, x0=NULL, x1=x0, y=NULL, ymax=NULL, ymin=NULL, r0=NULL, r1=NULL, data.panel=1, colors=c("blue", "white", "yellow"), ...)
}
\arguments{
\item{karyoplot}{(a \code{KaryoPlot} object) This is the first argument to all data plotting functions of \code{karyoploteR}. A KaryoPlot object referring to the currently active plot.}

\item{data}{(a \code{GRanges}) A GRanges object with the data. If \code{data} is present, \code{chr} will be set to \code{seqnames(data)} and \code{x} to the midpoints of the rages in data. If no parameter \code{y} is specified and \code{data} has a column named \code{y} or \code{value} this column will be used to define the \code{y} value of each data point. (defaults to NULL)}

\item{chr}{(a charecter vector) A vector of chromosome names specifying the chromosomes of the data points. If \code{data} is not NULL, \code{chr} is ignored. (defaults to NULL)}

\item{y}{(a numeric vector) A numeric vector with the values of the data points. If \code{y} is not NULL, it is used instead of any data column in \code{data}. (defaults to NULL)}

\item{ymax}{(numeric) The maximum value of \code{y} to be plotted. If NULL, it is set to the max value of the selected data panel. (defaults to NULL)}

\item{ymin}{(numeric) The minimum value of \code{y} to be plotted. If NULL, it is set to the min value of the selected data panel. (defaults to NULL)}

\item{r0}{(numeric) r0 and r1 define the vertical range of the data panel to be used to draw this plot. They can be used to split the data panel in different vertical ranges (similar to tracks in a genome browser) to plot differents data. If NULL, they are set to the min and max of the data panel, it is, to use all the available space. (defaults to NULL)}

\item{r1}{(numeric) r0 and r1 define the vertical range of the data panel to be used to draw this plot. They can be used to split the data panel in different vertical ranges (similar to tracks in a genome browser) to plot differents data. If NULL, they are set to the min and max of the data panel, it is, to use all the available space. (defaults to NULL)}

\item{data.panel}{(numeric) The identifier of the data panel where the data is to be plotted. The available data panels depend on the plot type selected in the call to \code{\link{plotKaryotype}}. (defaults to 1)}

\item{colors}{(colors) A set of color used to determine the color associated with each value. Internally, it uses \code{\link[grDevices]{colorRamp}}. (defaults to c("blue", "white", "yellow"))}

\item{...}{The ellipsis operator can be used to specify any additional graphical parameters. Any additional parameter will be passed to the internal calls to the R base plotting functions.}
}
\value{
Returns the original karyoplot object, unchanged.
}
\description{
Plots the given data as a heatmap along the genome
}
\details{
Given regions of the genome with a start, end and a value, draws a heatmap-like representation, with the 
color of the region determined by its value. It is important to note that \code{kpHeatmap} will not extend
the regions in any way, so if regions are not contiguous, they will appear as a series of rectangles and
not as a continuous plot.
}
\examples{



set.seed(1000)
windows <- unlist(tileGenome(tilewidth=1000000, seqlengths = seqlengths(Hsapiens)))
y <- sin(x=c(1:length(windows))/10)

kp <- plotKaryotype("hg19", plot.type=1, chromosomes=c("chr1", "chr2"))

kpLines(kp, windows, y=y, r0=0.4, r1=0.6, ymin=-1, ymax=1)
kpAxis(kp, r0=0.4, r1=0.6, ymin=-1, ymax=1, cex=0.5)

kpHeatmap(kp, windows, y=y, colors = c("red", "black", "green"), r0=0, r1=0.2)
kpHeatmap(kp, windows, y=y, colors = c("green", "black", "red"), r0=0.2, r1=0.4)

#or we can provide all data into a single GRanges object
mcols(windows) <- data.frame(y=y)

kpHeatmap(kp, windows, r0=0.6, r1=0.8)
#non-contiguous regions appear as solitary rectangles
kpHeatmap(kp, sample(x = windows, 500), r0=0.8, r1=1, color=c("orange", "black", "purple", "green"))


}
\seealso{
\code{\link{plotKaryotype}}, \code{\link{kpRect}}, \code{\link{kpLines}}
}

