% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{lighter}
\alias{lighter}
\title{lighter}
\usage{
lighter(col, amount=150)
}
\arguments{
\item{col}{(color) The original color. Might be specified as a color name or a "#RRGGBB(AA)" hex color definition.}

\item{amount}{(integer, [0-255]) The fixed amount to add to each RGB channel (Defaults to 150).}
}
\value{
A lighter color
}
\description{
Given a color, return a lighter one
}
\details{
Very simple utility function to create lighter colors. Given a color, it
transforms it to rgb space, adds a set amount to all chanels and transforms
it back to a color.
}
\examples{
 
lighter("red")
lighter("#333333")
 
}
\seealso{
\code{\link{darker}}
}
