% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytobands.R
\name{plotCytobands}
\alias{plotCytobands}
\title{plotCytobands}
\usage{
plotCytobands(karyoplot, color.table=NULL, add.cytobands.names=FALSE, add.base.numbers=FALSE, ...)
}
\arguments{
\item{karyoplot}{a \code{karyoplot} object returned by a call to \code{plotKaryotype}}

\item{color.table}{(named character vector) a table specifying the colors to plot the cytobands. If NULL, it gets the colors calling \code{getCytobandColors}. (defaults to NULL)}

\item{add.cytobands.names}{(boolean) whether to add or not the cytoband names to the plot. (defaults to FALSE)}

\item{add.base.numbers}{(boolean) whether to add the base numbers to the plot. (defaults to FALSE)}

\item{...}{any additional parameter to be passed to the functions called from plotCytobands.}
}
\value{
invisibly returns the given karyoplot object
}
\description{
Plots the chromosome cytobands in a karyoplot
}
\details{
Plots the cytobands representing the chromosome structure in a karyoplot. It extracts the 
cytobands from the \code{karyoplot} object it recieves as a parameter. It is possible to 
specify the colors used to plot the cytobands. In addition, it can add the cytonand names
and base numbers to the plot.
}
\note{
In general, this function is automatically called by plotKaryotype
and the user never nees to call it.
}
\examples{


kp <- plotKaryotype(ideogram.plotter = NULL)
plotCytobands(kp)
 
}
\seealso{
\code{\link{plotKaryotype}}, \code{\link{getCytobandColors}}
}

