% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKaryotype.R
\name{plotKaryotype}
\alias{plotKaryotype}
\title{plotKaryotype}
\usage{
plotKaryotype(genome="hg19", ...)
}
\arguments{
\item{genome}{}
}
\value{
The coordinates change function to be used to plot additional data onto the karyotype
}
\description{
Create a new plot with a karyotype and nothing more and return the coordinates change function
}
\details{
Explicar els tipus de plots disponibles!
}
\examples{
 
}
\seealso{
\code{\link{getMask}}, \code{\link{getGenomeAndMask}}, \code{\link{characterToBSGenome}}, \code{\link{maskFromBSGenome}}, \code{\link{emptyCacheRegioneR}}
}

