% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareParameters4.R
\name{prepareParameters4}
\alias{prepareParameters4}
\title{prepareParameters4}
\usage{
prepareParameters4(function.name, karyoplot, data, chr, x0, x1, y0, y1, ymax, ymin, r0, r1, data.panel, ...)
}
\arguments{
\item{function.name}{(character) The name of the function calling \code{prepareParameters4}. Only user for error reporting.}

\item{karyoplot}{(KaryoPlot) A karyoplot object.}

\item{data}{A GRanges}

\item{chr}{A character representing the chromosome names.}

\item{x0}{The position in the chromosome in number of bases.}

\item{x1}{The position in the chromosome in number of bases.}

\item{y0}{The value to be plotted.}

\item{y1}{The value to be plotted.}

\item{ymax}{The maximum value of y}

\item{ymin}{The minimum value of y}

\item{r0}{The start of the range to use for plotting}

\item{r1}{The end of the range to use for plotting}

\item{data.panel}{The data panel to use}

\item{...}{Any additional parameter}
}
\value{
A list with five values: \code{chr}, \code{x0}, \code{x1}, \code{y0} and \code{y1}. Each of them 
a vector of the same length with the normalized values to plot.
}
\description{
Prepare and normalize the parameters for functions with x0, x1 and y0, y1 parameters
}
\details{
This function prepares and normalizes the parameters for plotting functions  
with x0, x1, y0 and y1 parameters (as opposed to x and y) so functions can
offer a richer interface while internally dealing only with standard and 
simple code. It extracts the 
positions from \code{data} if available and applies the \code{r0} and 
\code{r1} scaling. It returns the ready to plot values in a list with
only \code{chr}, \code{x0}, \code{x1}, \code{y0} and \code{y1}. 
All parameters are interpreted and used as explained in \code{kpRect}.
}
\note{
This function is only useful when creating custom plotting functions. 
It is not intended to the general user.

For detailed documentation on the parameters, see \code{\link{kpRect}}
}
\examples{

prepareParameters4("TestFunc", kp, data=NULL, chr="chr1", x0=c(10, 20, 30), y=c(0, 1, 2), r0=0, r1=0.5, ymin=0, ymax=2)

 
}
\seealso{
\code{\link{kpRect}}
}
