% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processAutotrack}
\alias{processAutotrack}
\title{processAutotrack}
\usage{
processAutotrack(r0, r1, autotrack)
}
\arguments{
\item{r0}{(numeric) the original r0}

\item{r1}{(numeric) the original r1}

\item{autotrack}{(list) the autotrack definition (current track(s), total tracks and margin)}
}
\value{
The adjusted r0 and r1 values to be used for plotting
}
\description{
Adjusts the r0 and r1 taking in to account the autotrack parameter
}
\details{
Simple utility function to adjust the r0 and r1 values given a valid
autotrack definition.
}
\examples{
 
processAutotrack(0, 1, list(1, 4, 0))
processAutotrack(0.5, 1, list(1, 4, 0))

processAutotrack(0, 1, list(c(1,3), 4, 0))

processAutotrack(0, 1, list(c(1,3), 4, 0.5))
 
}
