% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{transparent}
\alias{transparent}
\title{transparent}
\usage{
transparent(col, amount=150)
}
\arguments{
\item{col}{(color) The original color. Might be specified as a color name or a "#RRGGBB(AA)" hex color definition.}

\item{amount}{(number, [0-1]) The amount of transparency. 0 for completely visible, 1 for completely transparent. (Defaults to 0.5).}
}
\value{
A transparent color
}
\description{
Given a color, return a transparent one
}
\details{
Very simple utility function to create transparent colors. Given a color, it
transforms it to rgb space, adds a set amount to all chanels and transforms
it back to a color.
}
\examples{
 
transparent("red")
transparent("#333333")
 
}
\seealso{
\code{\link{lighter}}
}
